+++ src/common/health/health.controller.spec.ts
@@ -1,0 +1,65 @@
+import { Test, TestingModule } from "@nestjs/testing";
+import { HealthController } from "./health.controller";
+import {
+  HealthCheckService,
+  TypeOrmHealthIndicator,
+  MemoryHealthIndicator,
+  DiskHealthIndicator,
+} from "@nestjs/terminus";
+import { ConfigService } from "@nestjs/config";
+
+describe("HealthController", () => {
+  let controller: HealthController;
+  let healthService: HealthCheckService;
+
+  beforeEach(async () => {
+    const module: TestingModule = await Test.createTestingModule({
+      controllers: [HealthController],
+      providers: [
+        {
+          provide: HealthCheckService,
+          useValue: {
+            check: jest.fn(),
+          },
+        },
+        {
+          provide: TypeOrmHealthIndicator,
+          useValue: {
+            pingCheck: jest.fn(),
+          },
+        },
+        {
+          provide: MemoryHealthIndicator,
+          useValue: {
+            checkHeap: jest.fn(),
+            checkRSS: jest.fn(),
+          },
+        },
+        {
+          provide: DiskHealthIndicator,
+          useValue: {
+            checkStorage: jest.fn(),
+          },
+        },
+        {
+          provide: ConfigService,
+          useValue: {
+            get: jest.fn().mockReturnValue("0.2.4"),
+          },
+        },
+      ],
+    }).compile();
+
+    controller = module.get<HealthController>(HealthController);
+    healthService = module.get<HealthCheckService>(HealthCheckService);
+  });
+
+  it("should be defined", () => {
+    expect(controller).toBeDefined();
+  });
+
+  it("should call healthService.check", async () => {
+    await controller.check();
+    expect(healthService.check).toHaveBeenCalled();
+  });
+});